requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
google-auth==2.6.0
google-auth-oauthlib==0.5.1
google-auth-httplib2==0.1.0
google-api-python-client==2.40.0
python-dateutil==2.8.2